# m3: Speed up calculateSize and reduce allocations