calculateSize uses a custom TTransport to count the number of
bytes without moving bytes around. However, since it's not a
TRichTransport, when used with TCompactProtocol, string field
writes cast to `[]byte`, and use the `Write` method.

By modifying TCalcTransport to implement TRichTransport, we can
avoid these allocations and speed up the calculation.

BenchmarkCalulateSize (no tags) comparison:
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkCalulateSize-12     547           324           -40.77%

benchmark                    old allocs     new allocs     delta
BenchmarkCalulateSize-12     9              0              -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkCalulateSize-12     16            0             -100.00%
```

BenchmarkCalulateSize (1 tag) comparison:
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkCalulateSize-12     1098          561           -48.91%

benchmark                    old allocs     new allocs     delta
BenchmarkCalulateSize-12     16             0              -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkCalulateSize-12     56            0             -100.00%
```

See internal task T4903325 for more details.